🔥 [FEAT] : 글 유형 선택 handler 추가 및 파일 분리
