+++ src/views/Write/index.tsx
@@ -1,12 +1,20 @@
-import { HTMLAttributes, TextareaHTMLAttributes, useCallback, useEffect, useState } from 'react'
+import {
+  ButtonHTMLAttributes,
+  HTMLAttributes,
+  MouseEventHandler,
+  TextareaHTMLAttributes,
+  useEffect,
+  useState,
+} from 'react'
 import styled from 'styled-components'
 import { fontSize, FontSizeProps } from 'styled-system'
 import axios from 'axios'
 import { getWill, insertWill } from 'api/will'
 import { ArrowLeft } from 'components/Common/Svg'
 import { useRouter } from 'next/router'
-import { useModal } from 'components/Common'
-
+import { ModalCloseButton, useModal } from 'components/Common'
+import { Handler } from 'components/Common/Modal/types'
+import SelectPostTypeModal from 'views/Write/components/SelectPostTypeModal'
 const questionList = [
   '1. 살아오면서 가장 기뻤던 일은?',
   '2. 살아오면서 가장 자부심을 느꼈던 일은?',
@@ -20,35 +28,20 @@ const questionList = [
   '10. 살아오면서 가장 사랑했던 사람은?',
   '11. 죽음을 앞두고 가장 하고 싶은 말은?',
 ]
-const SelectPostTypeModal = (props) => {
-  const { handleClick } = props
-  console.log(handleClick)
-  return (
-    <St.SelectPostTypeModal>
-      <St.ModalTitle>유서 작성 방식을 선택할 수 있어요.</St.ModalTitle>
 
-      <St.ModalDescription>
-        유서를 처음 작성하시는 분들을 위해 두 가지의 선택방식을 두었어요. 편하신 방법을 선택하여, 당신의 마지막 일기를
-        작성해주세요.
-      </St.ModalDescription>
-      <St.ModalButton onClick={handleClick}>질문에 따라 유서를 적고 싶어요</St.ModalButton>
-      <St.ModalButton variant="primary">제 마음대로 유서를 적고 싶어요</St.ModalButton>
-    </St.SelectPostTypeModal>
-  )
-}
 const Write = () => {
   const [title, setTitle] = useState('')
   const [contents, setContents] = useState('')
   const router = useRouter()
   const [isDefaultPostType, setPostType] = useState(true)
-  const handlePostType = (e) => {
-    console.log('isFalse')
+  const handlePostType = () => {
     setPostType(false)
+    onDismiss()
   }
-  const [modal] = useModal(<SelectPostTypeModal handleClick={handlePostType} />)
+  const [modal, onDismiss] = useModal(<SelectPostTypeModal onClick={handlePostType} />)
   useEffect(() => {
     modal()
-  }, [isDefaultPostType])
+  }, [])
 
   const handleTitle = (e) => {
     setTitle(e.target.value)
@@ -107,11 +100,11 @@ const Write = () => {
       {isDefaultPostType ? (
         <Contents value={contents} onChange={handleContents}></Contents>
       ) : (
-        questionList.map((question) => (
-          <>
+        questionList.map((question, i) => (
+          <div key={`${i}-${question}`}>
             <div>{question}</div>
             <Contents value={contents} onChange={handleContents}></Contents>
-          </>
+          </div>
         ))
       )}
     </article>
@@ -126,56 +119,7 @@ const Contents = ({ ...props }: TextareaHTMLAttributes<HTMLTextAreaElement>) => 
 }
 
 interface TextAreaProps extends FontSizeProps, HTMLAttributes<HTMLTextAreaElement> {}
-type variant = 'primary' | 'secondary'
 const St = {
-  ModalTitle: styled.div`
-    font-weight: 600;
-    font-size: 18px;
-    margin-top: 32px;
-  `,
-  ModalDescription: styled.div`
-    width: 410px;
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 20px;
-    margin-top: 8px;
-    text-align: center;
-  `,
-  SelectPostTypeModal: styled.div`
-    width: 586px;
-    height: 375px;
-    background: #ffffff;
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.06), 0px 0px 1px rgba(0, 0, 0, 0.08);
-    border-radius: 4px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  `,
-  ModalButton: styled.button<{ variant?: variant }>`
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 14px 16px;
-    gap: 10px;
-
-    width: 335px;
-    height: 50px;
-    margin-top: 40px;
-    /* Color/Grayscale 2 */
-    background-color: ${({ variant, theme }) =>
-      variant === 'primary' ? theme.colors.grayscale7 : theme.colors.grayscale2};
-    color: ${({ variant, theme }) => (variant === 'primary' ? theme.colors.grayscale0 : theme.colors.grayscale7)};
-
-    border-radius: 4px;
-    border: none;
-
-    font-family: 'SUIT';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 18px;
-    line-height: 22px;
-  `,
   MenuBar: styled.div`
     border: none;
     height: 78px;
